Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,8 +6,8 @@ int main() {
     int o,j;
     scanf("%d %d %c",&a,&b,&c);
     if( c == '+') {
-        int u = a+b;
-        printf("%d",u);
+        float u = a+b;
+        printf("%f",u);
     } else if (c == "-") {
         int u = a-b;
          printf("%d",u);
